Fix CI mode exit code for PATH conflicts
When running in CI mode, don't return error code 1 when PATH conflicts
are detected. The installation and default setting were successful, and
PATH conflicts are just warnings that can't be fixed interactively in CI.

This fixes the test failure for 'Install with --default flag' where
the command was succeeding but exiting with code 1 due to PATH warnings.